refactor(pns): one probe set per event, and an engine that only decides
Findings 5, 1 and 3.

An event built TWO SystemProbes, one of them with an empty marker path, so
the engine and the banner each took their own idle and focus readings
milliseconds apart and could disagree about the same event. There is one
probe set now, shared by reference; four blanket impls make a reference
satisfy what its owner satisfies, which is what lets them share without
anyone owning it twice.

decide spanned 121 lines and did six jobs. The presence arbitration moves to
wants_the_phone, where the comment about guarding every reading finally
describes the whole function, and decide is left assembling the route and the
pane result.

The engine also owned two things that were never decisions: config fallback
policy, which now sits beside the registry it selects from, and path
defaulting, which sits at the composition root that builds every path it is
asked about. Their tests moved with them.
